Fix implicit float to int conversion
The value of "$sub->$paramToDecide" can be amount with
the default type of float. If float was used as an array
key, it was implicitly type casted to integer and some
precision was lost silently.

The array_key_exists call already type-casted the value
to string, this commit expands this change and uses string
value everywhere where necessary.

remp/crm#2706